+++ user/cli/CLIDeleteUserCommand.class.php
@@ -3,7 +3,7 @@
  *                          CLIDeleteUserCommand.class.php
  *                            -------------------
  *   begin                : October 11, 2011
- *   copyright            : (C) 2011 Kévin MASSY
+ *   copyright            : (C) 2011 Kï¿½vin MASSY
  *   email                : soldier.weasel@gmail.com
  *
  *
@@ -65,9 +65,9 @@ class CLIDeleteUserCommand implements CLICommand
 		if (!empty($this->id))
 		{
 			$this->show_parameter('--id', $this->id);
-			if(UserService::user_exists_by_id($this->id))
+			if(UserService::user_exists('WHERE user_id=:user_id', array('user_id' => $this->id)))
 			{
-				UserService::delete_account($this->id);
+				UserService::delete_account('WHERE user_id=:user_id', array('user_id' => $this->id));
 				$this->write_success_message();
 			}
 			else
@@ -79,9 +79,9 @@ class CLIDeleteUserCommand implements CLICommand
 		else if (!empty($this->login))
 		{
 			$this->show_parameter('--login', $this->login);
-			if(UserService::user_exists_by_login($this->login))
+			if(UserService::user_exists('WHERE login=:login', array('login' => $this->login)))
 			{
-				UserService::delete_account_by_login($this->login);
+				UserService::delete_account('WHERE login=:login', array('login' => $this->login));
 				$this->write_success_message();
 			}
 			else
@@ -93,9 +93,9 @@ class CLIDeleteUserCommand implements CLICommand
 		else if (!empty($this->email))
 		{
 			$this->show_parameter('--email', $this->email);
-			if(UserService::user_exists_by_email($this->email))
+			if(UserService::user_exists('WHERE user_mail=:email', array('email' => $this->email)))
 			{
-				UserService::delete_account_by_email($this->email);
+				UserService::delete_account('WHERE user_mail=:email', array('email' => $this->email));
 				$this->write_success_message();
 			}
 			else
